requirements.txt
pytermtk
setuptools
python-ulid
requests
cryptography
